Give the renderer a reference to the builder
The builder keeps usefult config settings like the package
being documented and the directory where the generated qmd files
are placed.

The builder also keeps track of the documented items that are
later converted into the inventory on which interlinking depends.

With this change the renderers have access to this information
as well.